# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "druthers.io API local",
    "description": "API for druthers.io \u2014 track and rank the movies, TV, books, and games you actually care about.",
    "contact": {
      "name": "Adam",
      "url": "https://www.druthers.io/",
      "email": "aleonard9@hotmail.com"
    },
    "license": {
      "name": "GPL-3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/v1/users/me/visibility": {
      "get": {
        "tags": [
          "Visibility"
        ],
        "summary": "Get Visibility",
        "operationId": "get_visibility_v1_users_me_visibility_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutVisibility"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Visibility"
        ],
        "summary": "Update Visibility",
        "description": "Update the handle and/or per-category public flags. Opening any category\nto the public requires a handle (that's the profile URL).",
        "operationId": "update_visibility_v1_users_me_visibility_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InVisibilityUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutVisibility"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/public/{handle}": {
      "get": {
        "tags": [
          "Visibility"
        ],
        "summary": "Public Profile",
        "description": "Read-only public profile: ranked lists of the categories the owner has\nopted in, best first. Fully private profiles (and unknown handles) 404\nidentically, so the endpoint never confirms a private account exists.",
        "operationId": "public_profile_v1_public__handle__get",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Handle"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/me/summary": {
      "get": {
        "tags": [
          "Summary"
        ],
        "summary": "Get Summary",
        "description": "Per-shelf Top 5 plus ranked/queued counts for the signed-in user.\n\nReplaces the home page's four full-collection fetches; row count is\nindependent of library size.",
        "operationId": "get_summary_v1_users_me_summary_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "top",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5,
              "minimum": 1,
              "description": "Entries per shelf",
              "default": 5,
              "title": "Top"
            },
            "description": "Entries per shelf"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InVisibilityUpdate": {
        "properties": {
          "handle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handle"
          },
          "public_movies": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Movies"
          },
          "public_tv": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Tv"
          },
          "public_books": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Books"
          },
          "public_games": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Games"
          }
        },
        "type": "object",
        "title": "InVisibilityUpdate",
        "description": "Request body for visibility settings. Only sent fields change; a null\nhandle clears it (allowed only while everything is private)."
      },
...
      "OutSummary": {
        "properties": {
          "handle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handle"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "profile_public": {
            "type": "boolean",
            "title": "Profile Public",
            "default": false
          },
          "shelves": {
            "items": {
              "$ref": "#/components/schemas/OutSummaryShelf"
            },
            "type": "array",
            "title": "Shelves"
          },
          "total_ranked": {
            "type": "integer",
            "title": "Total Ranked"
          }
        },
        "type": "object",
        "required": [
          "shelves",
          "total_ranked"
        ],
        "title": "OutSummary",
        "description": "Everything the home page renders, in one bounded response \u2014 see\napp/services/summary.py for why this endpoint exists."
      },
      "OutSummaryEntry": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "id",
          "title"
        ],
        "title": "OutSummaryEntry",
        "description": "One ranked entry on a shelf's Top 5."
      },
      "OutSummaryShelf": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "ranked_count": {
            "type": "integer",
            "title": "Ranked Count"
          },
          "queued_count": {
            "type": "integer",
            "title": "Queued Count"
          },
          "public": {
            "type": "boolean",
            "title": "Public"
          },
          "top": {
            "items": {
              "$ref": "#/components/schemas/OutSummaryEntry"
            },
            "type": "array",
            "title": "Top"
          }
        },
        "type": "object",
        "required": [
          "category",
          "label",
          "ranked_count",
          "queued_count",
          "public",
          "top"
        ],
        "title": "OutSummaryShelf",
        "description": "One domain's headline numbers plus its best-ranked entries."
      },
...
      "OutVisibility": {
        "properties": {
          "handle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handle"
          },
          "public_movies": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Movies",
            "default": false
          },
          "public_tv": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Tv",
            "default": false
          },
          "public_books": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Books",
            "default": false
          },
          "public_games": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Games",
            "default": false
          }
        },
        "type": "object",
        "title": "OutVisibility",
        "description": "The caller's visibility settings. NULL flags read as private."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "v1/auth/token"
          }
        }
      }
    }
  },
...
}